requirements/default.txt
django-ordered-model>=3.0
django-colorfield==0.4.1
